✨ Feature: DB 연동화

+++ src/main/java/site/dataon/hyeyum/domain/CompanyFinancialStatistics.java
@@ -44,19 +44,19 @@ public class CompanyFinancialStatistics {
 
     @Comment("매출액//예시)  28,018,841")
     @Column(name = "sales_amount")
-    private Integer salesAmount;
+    private Long salesAmount;
 
     @Comment("영업이익손실//예시) -540,086")
     @Column(name = "operating_income")
-    private Integer operatingIncome;
+    private Long operatingIncome;
 
     @Comment("매출원가//예시)  24,399,948")
     @Column(name = "cost_of_sales")
-    private Integer costOfSales;
+    private Long costOfSales;
 
     @Comment("당기순이익손실//예시) -2,820,077")
     @Column(name = "net_income")
-    private Integer netIncome;
+    private Long netIncome;
 
     @Comment("영업이익률//예시) -53.17")
     @Column(name = "operating_margin")
@@ -64,23 +64,23 @@ public class CompanyFinancialStatistics {
 
     @Comment("자산총계//예시)  4,281,095")
     @Column(name = "total_assets")
-    private Integer totalAssets;
+    private Long totalAssets;
 
     @Comment("부채총계//예시)  8,355,775")
     @Column(name = "total_liabilities")
-    private Integer totalLiabilities;
+    private Long totalLiabilities;
 
     @Comment("자본총계//예시)  205,024")
     @Column(name = "total_equity")
-    private Integer totalEquity;
+    private Long totalEquity;
 
     @Comment("납입자본금//예시)  950,000")
     @Column(name = "paid_in_capital")
-    private Integer paidInCapital;
+    private Long paidInCapital;
 
     @Comment("연구개발비//예시)  773,482")
     @Column(name = "research_and_development_expense")
-    private Integer researchAndDevelopmentExpense;
+    private Long researchAndDevelopmentExpense;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "company_id", insertable = false, updatable = false)
